+++ projects/fhi-angular-highcharts/src/lib/diagram-types/fhi-diagram-types.ts
