+++ projects/fhi-angular-highcharts/src/lib/highcharts-options/options-map-fylker.ts
@@ -1,4 +1,4 @@
-import { DiagramTypeId } from '../fhi-diagram-types/fhi-diagram-type-id';
+import { DiagramTypeId } from '../diagram-types/fhi-diagram-type-id';
 
 export const OptionsMapFylker = {
   chart: {
